# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Future Card API - Digital memory cards for children
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CardStatus = typeof CardStatus[keyof typeof CardStatus];


export const CardStatus = {
  draft: 'draft',
  generating: 'generating',
  completed: 'completed',
} as const;

export type CardLanguage = typeof CardLanguage[keyof typeof CardLanguage];


export const CardLanguage = {
  en: 'en',
  ar: 'ar',
} as const;

export interface Card {
  id: string;
  childName: string;
  profession: string;
  /** @nullable */
  customProfession?: string | null;
  /** @nullable */
  childPhotoUrl?: string | null;
  /** @nullable */
  aiImageUrl?: string | null;
  /** @nullable */
  voiceMessageUrl?: string | null;
  /** @nullable */
  parentMessage?: string | null;
  /** @nullable */
  qrCodeUrl?: string | null;
  status: CardStatus;
  language?: CardLanguage;
  createdAt: string;
  updatedAt: string;
}

export type CardInputLanguage = typeof CardInputLanguage[keyof typeof CardInputLanguage];


export const CardInputLanguage = {
  en: 'en',
  ar: 'ar',
} as const;

export interface CardInput {
  /** @minLength 1 */
  childName: string;
  /** @minLength 1 */
  profession?: string;
  customProfession?: string;
  childPhotoUrl?: string;
  voiceMessageUrl?: string;
  parentMessage?: string;
  language?: CardInputLanguage;
}

export type CardUpdateStatus = typeof CardUpdateStatus[keyof typeof CardUpdateStatus];


export const CardUpdateStatus = {
  draft: 'draft',
  generating: 'generating',
  completed: 'completed',
} as const;

export type CardUpdateLanguage = typeof CardUpdateLanguage[keyof typeof CardUpdateLanguage];


export const CardUpdateLanguage = {
  en: 'en',
  ar: 'ar',
} as const;

export interface CardUpdate {
  /** @minLength 1 */
  childName?: string;
  profession?: string;
  customProfession?: string;
  childPhotoUrl?: string;
  aiImageUrl?: string;
  voiceMessageUrl?: string;
  parentMessage?: string;
  qrCodeUrl?: string;
  status?: CardUpdateStatus;
  language?: CardUpdateLanguage;
}

export interface ProfessionCount {
  profession: string;
  count: number;
}

export interface CardStats {
  total: number;
  completed: number;
  drafts: number;
  generating: number;
  professionBreakdown: ProfessionCount[];
  recentCards: Card[];
}

export interface PublicCard {
  id: string;
  childName: string;
  profession: string;
  /** @nullable */
  customProfession?: string | null;
  /** @nullable */
  aiImageUrl?: string | null;
  /** @nullable */
  childPhotoUrl?: string | null;
  /** @nullable */
  voiceMessageUrl?: string | null;
  /** @nullable */
  parentMessage?: string | null;
  language?: string;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListCardsParams = {
/**
 * Search by child name or profession
 */
search?: string;
};