# lib/api-zod/src/generated/types/cardInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Future Card API - Digital memory cards for children
 * OpenAPI spec version: 0.1.0
 */
import type { CardInputLanguage } from './cardInputLanguage';

export interface CardInput {
  /** @minLength 1 */
  childName: string;
  /** @minLength 1 */
  profession?: string;
  customProfession?: string;
  childPhotoUrl?: string;
  voiceMessageUrl?: string;
  parentMessage?: string;
  language?: CardInputLanguage;
}